+++ src/preload/index.ts
@@ -170,10 +170,8 @@ const api = {
     ipcRenderer.on('tools-download-details', handler)
     return () => ipcRenderer.removeListener('tools-download-details', handler)
   },
-  onDllUrlRequired: (callback: () => void) => {
-    ipcRenderer.on('dll-url-required', callback)
-    return () => ipcRenderer.removeListener('dll-url-required', callback)
-  },
+  checkDllExist: () => ipcRenderer.invoke('check-dll-exist'),
+  openToolsFolder: () => ipcRenderer.invoke('open-tools-folder'),
 
   // Window controls
   minimizeWindow: () => ipcRenderer.send('window-minimize'),